Gracefully truncate search result message if too long.
Sonica will now check if an addition to a message would make the message too long, and will insert an 'overflow' message if so.
Also added some packages to discord and daemon requirements file, as these packages were missing, and had only been added in default.nix for each module.